Proverbs_Law_MainPage_Official/requirements.txt
gradio==6.2.0
huggingface-hub>=0.20.0
transformers>=4.35.0
torch>=2.0.0
pillow>=10.0.0
datasets>=2.15.0
PyPDF2>=3.0.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
python-docx>=0.8.11
numpy>=1.24.0
requests>=2.31.0